+++ lib/src/public/core/channel/group_channel/mfm/uploaded_file_info.dart
@@ -46,6 +46,7 @@ class UploadedFileInfo {
 
   bool requireAuth;
 
+  @JsonKey(includeFromJson: false, includeToJson: false)
   File? file;
 
   UploadedFileInfo({