+++ lib/src/public/main/params/message/file_message_create_params.dart
@@ -17,6 +17,10 @@ class FileMessageCreateParams extends BaseMessageCreateParams {
   /// The file's thumbnail sizes of the messages. Defaults to null.
   List<Size>? thumbnailSizes;
 
+  /// The message text.
+  /// @since 4.9.0
+  String? message;
+
   /// withFile
   FileMessageCreateParams.withFile(
     File file, {
@@ -31,6 +35,7 @@ class FileMessageCreateParams extends BaseMessageCreateParams {
     PushNotificationDeliveryOption pushNotificationDeliveryOption =
         PushNotificationDeliveryOption.normal,
     bool isPinnedMessage = false,
+    this.message,
   }) : super(
           data: data,
           customType: customType,
@@ -98,6 +103,7 @@ class FileMessageCreateParams extends BaseMessageCreateParams {
     PushNotificationDeliveryOption pushNotificationDeliveryOption =
         PushNotificationDeliveryOption.normal,
     bool isPinnedMessage = false,
+    this.message,
   }) : super(
           data: data,
           customType: customType,
@@ -152,6 +158,7 @@ class FileMessageCreateParams extends BaseMessageCreateParams {
     PushNotificationDeliveryOption pushNotificationDeliveryOption =
         PushNotificationDeliveryOption.normal,
     bool isPinnedMessage = false,
+    this.message,
   })  : fileInfo = createFileInfoFromFileUrl(
           fileName: fileName ?? 'file',
           mimeType: mimeType ?? 'application/octet-stream',
@@ -187,6 +194,8 @@ class FileMessageCreateParams extends BaseMessageCreateParams {
   /// withMessage
   FileMessageCreateParams.withMessage(FileMessage fileMessage)
       : super.withMessage(fileMessage) {
+    message = fileMessage.message;
+
     if (fileMessage.url.isNotEmpty) {
       fileInfo = FileInfo.fromFileUrl(
         fileUrl: fileMessage.url,
@@ -208,6 +217,7 @@ class FileMessageCreateParams extends BaseMessageCreateParams {
     ret['name'] = fileInfo.fileName;
     ret['size'] = fileInfo.fileSize;
     ret['type'] = fileInfo.mimeType;
+    ret['message'] = message;
 
     ret.removeWhere((key, value) => value == null);
     return ret;